+++ app/data/data.go
@@ -1,17 +1,12 @@
 package data
 
 import (
-	"context"
 	"encoding/hex"
 	"encoding/json"
 	"fmt"
 	"log"
-	"regexp"
 	"strings"
-	"time"
 
-	"github.com/go-redis/redis/v8"
-	"github.com/gorilla/websocket"
 	"github.com/lib/pq"
 )
 
@@ -185,166 +180,4 @@ func (e *Events) ToJSON() []byte {
 
 	return data
 
-}
-
-// SubscriptionRequest - Real time data subscription/ unsubscription request
-// needs to be sent in this form, from client application
-type SubscriptionRequest struct {
-	Name string `json:"name"`
-	Type string `json:"type"`
-}
-
-// IsValidTopic - Checks whether topic to which client application is trying to
-// subscribe to is valid one or not
-func (s *SubscriptionRequest) IsValidTopic() bool {
-	pattern, err := regexp.Compile("^(block|(transaction(/(0x[a-zA-Z0-9]{40}|\\*)(/(0x[a-zA-Z0-9]{40}|\\*))?)?))$")
-	if err != nil {
-		log.Printf("[!] Failed to parse regex pattern : %s\n", err.Error())
-		return false
-	}
-
-	if pattern.MatchString(s.Name) {
-
-		matches := pattern.FindStringSubmatch(s.Name)
-		if strings.HasPrefix(matches[0], "block") {
-			return true
-		} else if strings.HasPrefix(matches[0], "transaction") {
-			return true
-		}
-
-	}
-
-	return false
-}
-
-// Validate - Validates request from client for subscription/ unsubscription
-func (s *SubscriptionRequest) Validate(topics map[string]bool) bool {
-
-	// --- Closure definition
-	// Given associative array & key in array, checks whether entry exists or not
-	// If yes, also return entry's boolean value
-	checkEntryInAssociativeArray := func(name string) bool {
-		v, ok := topics[name]
-		if !ok {
-			return false
-		}
-
-		return v
-	}
-	// ---
-
-	var validated bool
-
-	switch s.Type {
-	case "subscribe":
-		validated = s.IsValidTopic() && !checkEntryInAssociativeArray(s.Name)
-	case "unsubscribe":
-		validated = s.IsValidTopic() && checkEntryInAssociativeArray(s.Name)
-	default:
-		validated = false
-	}
-
-	return validated
-
-}
-
-// SubscriptionResponse - Real time data subscription/ unsubscription request to be responded with
-// in this form
-type SubscriptionResponse struct {
-	Code    uint   `json:"code"`
-	Message string `json:"msg"`
-}
-
-// BlockConsumer - To be subscribed to `block` topic using this consumer handle
-// and client connected using websocket needs to be delivered this piece of data
-type BlockConsumer struct {
-	Client     *redis.Client
-	Channel    string
-	Connection *websocket.Conn
-	PubSub     *redis.PubSub
-}
-
-// Subscribe - Subscribe to `block` channel
-func (b *BlockConsumer) Subscribe() {
-	b.PubSub = b.Client.Subscribe(context.Background(), b.Channel)
-}
-
-// Listen - Listener function, which keeps looping in infinite loop
-// and reads data from subcribed channel, which also gets delivered to client application
-func (b *BlockConsumer) Listen() {
-
-	for {
-		msg, err := b.PubSub.ReceiveTimeout(context.Background(), time.Second)
-		if err != nil {
-			continue
-		}
-
-		switch m := msg.(type) {
-		case *redis.Subscription:
-
-		case *redis.Message:
-			if !b.Send(m.Payload) {
-				break
-			}
-		}
-	}
-
-}
-
-// Send - Tries to deliver subscribed block data to client application
-// connected over websocket
-func (b *BlockConsumer) Send(msg string) bool {
-	var block Block
-
-	_msg := []byte(msg)
-
-	err := json.Unmarshal(_msg, &block)
-	if err != nil {
-		log.Printf("[!] Failed to decode published block data to JSON : %s\n", err.Error())
-		return true
-	}
-
-	if err = b.Connection.WriteJSON(&block); err != nil {
-		log.Printf("[!] Failed to deliver block data to client : %s\n", err.Error())
-
-		if err = b.PubSub.Unsubscribe(context.Background(), b.Channel); err != nil {
-			log.Printf("[!] Failed to unsubscribe from block event : %s\n", err.Error())
-		}
-
-		if err = b.Connection.Close(); err != nil {
-			log.Printf("[!] Failed to close websocket connection : %s\n", err.Error())
-		}
-
-		return false
-	}
-
-	log.Printf("[!] Delivered block data to client\n")
-	return true
-}
-
-// SendConfirmation - Sending confirmation message i.e. block subscription has been confirmed
-// for client. If unable to send it, cancels subscription & closes underlying websocket connection
-//
-// Websocket connection may already be closed, in that case it'll simply return
-func (b *BlockConsumer) SendConfirmation() bool {
-
-	if err := b.Connection.WriteJSON(&SubscriptionResponse{
-		Code:    1,
-		Message: "Subscribed to `block`",
-	}); err != nil {
-		log.Printf("[!] Failed to deliver block subscription confirmation to client : %s\n", err.Error())
-
-		if err = b.PubSub.Unsubscribe(context.Background(), b.Channel); err != nil {
-			log.Printf("[!] Failed to unsubscribe from block event : %s\n", err.Error())
-		}
-
-		if err = b.Connection.Close(); err != nil {
-			log.Printf("[!] Failed to close websocket connection : %s\n", err.Error())
-		}
-
-		return false
-	}
-
-	log.Printf("[!] Delivered block subscription confirmation to client\n")
-	return true
 }
